+++ CHANGES.rst
@@ -1,3 +1,7 @@
+Next release
+---------------
+- [Bug] :issue:`109`: Fix automated session closure handled by python garbage collection
+
 1.6.3 (03/12/2020)
 ------------------
 - [Improvement]: remove pytest-runner from setup_requires as this is deprecated for security reasons, see https://github.com/pytest-dev/pytest-runner